refactor: component파일들 components 폴더로 이동 시킴
