--- app/modificationHistory/modificationHistory.module.scss
@@ -1,107 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-  background-color: var(--content);
-  border-radius: 12px;
-  padding: 20px 20px;
-  width: 100%;
-
-  .totalCount {
-    font-size: 1.7rem;
-    strong {
-      font-weight: bold;
-    }
-  }
-
-  .listWapper {
-    display: flex;
-    flex-direction: column;
-    border-top: 2px solid var(--foreground);
-    border-bottom: 2px solid var(--foreground);
-    width: 100%;
-    height: 624px;
-    overflow-y: auto;  
-
-    .listTitle {
-      height: 52px;
-    }
-
-    .listTitle, .listLine {
-      display: flex;
-      width: 100%;
-    }
-
-    div {
-      text-align: center;
-    }
-
-    .listLine div:nth-child(4) {
-      min-width: 0; 
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      text-align: left;
-    }
-
-    .listTitle {
-      font-size: 2.4rem;
-      border-bottom: 1px solid var(--foreground);
-      align-items: center;
-    }
-
-    .listContent {
-      display: flex;
-      flex-direction: column;
-      padding: 8px 0;
-      gap: 12px;
-      
-      .listLine {
-        font-size: 1.9rem;
-        background-color: var(--background);
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        height: 71.75px;
- 
-        div.status {
-          &.pending {
-            color: var(--secondary-2); // 대기
-          }
-
-          &.complete {
-            color: var(--foreground); // 정정 완료
-          }
-        }
-        
-      }
-  
-      .noData {
-        font-size: 1.9rem;
-        background-color: var(--background);
-        text-align: center;
-      }
-    }
-
-    .listTitle div:nth-child(1),
-    .listLine div:nth-child(1),
-    .listTitle div:nth-child(6),
-    .listLine div:nth-child(6) {
-      flex: 1;
-    }
-  
-    .listTitle div:nth-child(2),
-    .listLine div:nth-child(2),
-    .listTitle div:nth-child(3),
-    .listLine div:nth-child(3),
-    .listTitle div:nth-child(5),
-    .listLine div:nth-child(5) {
-        flex: 2;
-    }
-  
-    .listTitle div:nth-child(4),
-    .listLine div:nth-child(4) {
-        flex: 3;
-    }
-  }
-}
