+++ app/modificationHistory/page.jsx
@@ -1,5 +1,6 @@
 "use client";
-import ModificationHistory from "./(components)/modificationHistory";
+
+import ModificationHistory from "../../components/modificationHistory/modificationHistory";
 
 export default function ModificationHistoryPage() {
   return <ModificationHistory />;
